fix demo
