+++ src/website/src/pages/ec/components/timeline/docs/code.mdx
@@ -4,10 +4,7 @@ order: 2
 ---
 
 import Playground from '@ecl/website-components/Playground';
-
-import Timeline from '@ecl/ec-react-component-timeline/src/Timeline2';
-
-import demoContent from '@ecl/ec-specs-timeline/demo/data';
+import TimelineExample from '@ecl/ec-react-component-timeline/examples/Timeline2';
 
 <Playground
   system="ec"
@@ -15,5 +12,5 @@ import demoContent from '@ecl/ec-specs-timeline/demo/data';
   selectedStory="default"
   showFrame
 >
-  <Timeline {...demoContent} />
+  <TimelineExample />
 </Playground>
