+++ src/website/src/pages/eu/components/timeline/docs/code.mdx
@@ -4,10 +4,7 @@ order: 2
 ---
 
 import Playground from '@ecl/website-components/Playground';
-
-import Timeline from '@ecl/eu-react-component-timeline/src/Timeline2';
-
-import demoContent from '@ecl/eu-specs-timeline/demo/data';
+import TimelineExample from '@ecl/eu-react-component-timeline/examples/Timeline2';
 
 <Playground
   system="eu"
@@ -15,5 +12,5 @@ import demoContent from '@ecl/eu-specs-timeline/demo/data';
   selectedStory="default"
   showFrame
 >
-  <Timeline {...demoContent} />
+  <TimelineExample />
 </Playground>
